Switch to using whitenoise
Trying to deploy the develop branch to staging the assets are not being
served at all through Django. Heroku's documentation on Django assets
[1] suggests using a package called whitenoise [2] that serves the
assets directly through the WSGI server.

This change also pulls a lot of the settings (particularly the ones
related to static files) up into the base settings so that asset serving
can be more easily tested locally.

[1] https://devcenter.heroku.com/articles/django-assets
[2] http://whitenoise.evans.io