Improve Docker file and compose system
This fixes the order of asset building commands in the Dockerfile so
that collectstatic happens last (otherwise not all assets are
collected). It also improves the Dockerfile by not copying the whole
app directory early and therefore allowing the layer cache to be used
more effectively.

In the compose file the local run is made more like production so that
asset serving can be tested locally.