Remove the deployment of the app from Circle CI
Previously we were building assets and deploying the app using Circle
CI rather than via heroku.

This was because we also had Ruby as a dependency and we wanted to keep
the app neater.

Now that this dependency has been removed we've moved to allowing
heroku to build and deploy the app.