Support images within homepage panels
This is a temporary solution which doesn't involve changing wagtail
or the database to support images on the homepage.

It makes use of the classname field to populate the images. This
new panel idea is being tested first and if the decision is made to
keep it, it will need to be implemented correctly within the CMS.