Add an ID to each heading in content blocks
This change allows anchor links to be created to specific bits of
content within a long page.

At the moment there is no way to generate a link to a manual anchor
within the richtext editor.

This slugifys the text of a heading and uses this value as the ID.