Remove lodash from client JavaScript
Lodash was being used in a small number of modules to bind the
context of this to a method.

The whole lodash library was being included just to use this one
method which added a lot of bloat to the JavaScript (~400k unminified).

This change removes it from the client to reduce the size of the
clientside script.