Upgrade Node and NPM
The project was running a version of Node that was no longer in
active support.

This change upgrades to the latest long term support version, v10.

It also includes an update to a package that no longer worked with
the latest version of Node and an update to supress some warnings
about the package manifest.