Update jQuery
This site was still using jQuery 1 which flagged up some security
vulnerabilities around Cross Site Scripting:
https://nodesecurity.io/advisories/328

This updates to jQuery 3 which does drop some browser support.

However, after reviewing the analytics of the site this accounts for
a small proportion and JavaScript is not required for the site to
function so we can stop serving JavaScript to these older browsers.