Remove breadcrumb
After the wider navigation changes the breadcrumb no longer provides
as useful way to navigate.

The majority of pages are only one level deep so the breadcrumg only
provides a link back to the homepage, which has now been added to the
global navigation.

This seemed to cause duplication so for now it is being removed.