Update search form to show label
The search form was currently using a hidden label and showing the
label text within a placeholder. This means having a larger box
size which sometimes led to people entering things other than a
postcode.

This change moves the label outside and explains to enter the full
postcode as well as reducing the size of the box to suggest that
only a postcode should be entered.