Add support to select image within Wagtail
Previously we used a temporary solution to set the image for a panel
based on the classname within the CMS.

Long term we want to be able to choose that image from the Wagtail.

This change adds support for it within the edit interface and removes
the previous method of displaying panel images.